Added better error checks for apis. Did some changes to database tables (with a new user type Admin). Added a seeder for initial test data. Updated api routes.

+++ routes/api.php
@@ -5,6 +5,8 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\Api\UserController;
 use App\Http\Controllers\Api\MemberController;
 use App\Http\Controllers\Api\TrainerController;
+use App\Http\Controllers\Api\AdminController;
+use GuzzleHttp\Middleware;
 
 Route::get('/user', function (Request $request) {
     return $request->user();
@@ -13,40 +15,61 @@ Route::get('/user', function (Request $request) {
 // API Routes
 
 // Public Routes
-Route::post('/register', [UserController::class, 'register']);
-Route::post('/login', [UserController::class, 'login']);
+Route::post('/login', [UserController::class, 'login'])->name('user.login');
 
 // Protected Routes
 Route::group([
-    'middleware' => 'auth:api'
+    'middleware' => 'auth:api',
 ], function () {
 
-    // Common routes accessible by all authenticated users
-    Route::get('/user', [UserController::class, 'profile'])->name('user.profile');
-    Route::post('/refresh-token', [UserController::class, 'refreshToken'])->name('token.refresh');
-    Route::post('/logout', [UserController::class, 'logout'])->name('user.logout');
-    Route::put('/user', [UserController::class, 'updateUser'])->name('user.update');
-    Route::delete('/user', [UserController::class, 'deleteUser'])->name('user.delete');
+    // Common Auth Routes
+    Route::post('/logout', [UserController::class, 'logout'])->name('user.logout'); // Logout user
+    Route::post('/refresh-token', [UserController::class, 'refreshToken'])->name('token.refresh'); // Refresh token
 
-    // Member Routes
+    // Common User Routes
+    Route::get('/user/{userId}', [UserController::class, 'getUser'])->name('user.get'); // Get user
+    Route::put('/user/{userId}', [UserController::class, 'updateUser'])->name('user.update'); // Update user
+    Route::delete('/user/{userId}', [UserController::class, 'deleteUser'])->name('user.delete'); // Delete user
+
+    // Admin Routes (only for admins)
     Route::group(
-        [
-            'middleware' => 'member'
-        ],
+        ['middleware' => 'admin'],
         function () {
-            //  Route::post('/member/{userId}', [MemberController::class, 'createMember'])->name('member.create');
-            Route::put('/member/{userId}', [MemberController::class, 'updateMember'])->name('member.update');
+            Route::post('/register', [UserController::class, 'register'])->name('user.register'); // Create and Register new user
+            Route::get('/admin/{userId}', [AdminController::class, 'getAdmin'])->name('admin.get'); // Get admin
+            Route::get('/admins', [AdminController::class, 'getAllAdmins'])->name('admin.get.all'); // Get all admins
+            Route::get('/members', [MemberController::class, 'getAllMembers'])->name('member.get.all'); // Get all members
+            Route::get('/trainers', [TrainerController::class, 'getAllTrainers'])->name('trainer.get.all'); // Get all trainers
+
+            // automatic routes (backup routes) 
+            Route::post('/admin', [AdminController::class, 'createAdmin'])->name('admin.create'); // Create new admin
+            Route::post('/member', [MemberController::class, 'createMember'])->name('member.create'); // Create new member
+            Route::post('/trainer', [TrainerController::class, 'createTrainer'])->name('trainer.create'); // Create new trainer
+            Route::delete('/admin/{userId}', [AdminController::class, 'deleteAdmin'])->name('admin.delete'); // Delete admin
         }
     );
 
-    // Trainer Routes
+    // Member Routes (for members and admins)
     Route::group(
-        [
-            'middleware' => 'trainer'
-        ],
+        ['middleware' => ['member', 'admin']],
         function () {
-            Route::post('/trainer/{userId}', [TrainerController::class, 'createTrainer'])->name('trainer.create');
-            Route::put('/trainer/{userId}', [TrainerController::class, 'updateTrainer'])->name('trainer.update');
+            Route::get('/member/{userId}', [MemberController::class, 'getMember'])->name('member.get'); // Get member
+            Route::put('/member/{userId}', [MemberController::class, 'updateMember'])->name('member.update'); // Update member
+
+            // automatic routes (backup routes) 
+            Route::delete('/member/{userId}', [MemberController::class, 'deleteMember'])->name('member.delete'); // Delete member
+        }
+    );
+
+    // Trainer Routes (for trainers and admins)
+    Route::group(
+        ['middleware' => ['member', 'admin']],
+        function () {
+            Route::get('/trainer/{userId}', [TrainerController::class, 'getTrainer'])->name('trainer.get'); // Get trainer
+            Route::put('/trainer/{userId}', [TrainerController::class, 'updateTrainer'])->name('trainer.update'); // Update trainer
+
+            // automatic routes (backup routes) 
+            Route::delete('/trainer/{userId}', [TrainerController::class, 'deleteTrainer'])->name('trainer.delete'); // Delete trainer
         }
     );
 });
